requirements.txt
flask
pdfplumber
sentence-transformers
faiss-cpu
transformers
torch
diffusers
accelerate
safetensors
huggingface-hub
